+++ cloud/objstore/objcli/objaws/utilities_test.go
@@ -35,6 +35,9 @@ func TestHandleError(t *testing.T) {
 	err = handleError(ptr.To("bucket1"), ptr.To("key1"), &smithy.GenericAPIError{Code: "AccessDenied"})
 	require.ErrorIs(t, err, objerr.ErrUnauthorized)
 
+	err = handleError(ptr.To("bucket1"), ptr.To("key1"), &smithy.GenericAPIError{Code: "Forbidden"})
+	require.ErrorIs(t, err, objerr.ErrUnauthorized)
+
 	err = handleError(ptr.To("bucket1"), ptr.To("key1"), &s3types.NoSuchKey{})
 	require.ErrorAs(t, err, &notFound)
 	require.Equal(t, "key", notFound.Type)